Create version-specific serialization objects
We need to support both serializing and deserializing the metric config. Currently, Promitor relies on the model matching the yaml schema, and just using the standard serialization from YamlDotNet.

The problem with this is that if we introduce multiple config formats, we end up with a situation where the model objects used by the application don't match both (or either) of the config formats anymore, and so serialization won't work. Given that most of the tests currently rely on this serialization, it would end up with a pretty complicated change to sort this.

So that we can still generate the old `v1` format, what I've done is created separate serialization objects for both config versions. This means that if you need to write out a config in a certain format, you can use those objects to define the configuration.

NOTE: this commit doesn't compile but I want to add the files like this to make it easier to see what I'm doing as I make further changes. It can be rebased into future commits if it makes sense.