Implement builder pattern for v1 models
Renamed all the models in the v1 namespace to `<ModelName>Builder` and added a `Build()` method to each of them. The reason for this is to distinguish between the model objects used by the application, and the models used for serialization.

I've also added additional constructors to the resource model objects. The reason I did this was to make it easier for me to be sure that the builders were populating all of the properties on each object. I've left the default constructors in place.