Update runtime model to support multiple resources per metric (#666)
* Alter MetricDefinition to allow multiple resources

I've updated `MetricDefinition` to support multiple resources so that we can define a single prometheus metric that is used to store the values for multiple Azure resources.

In order to do this, I've changed `MetricDefinition` so that it doesn't describe an Azure resource anymore, and created `AzureResourceDefinition` to be the base class for each resource type. I've also created a `PrometheusMetricDefinition` class to contain the details of the prometheus metric being defined.

In addition, I've added a new class, `ScrapeDefinition`, that describes a single Azure resource being scraped. This is used in the places where `MetricDefinition` would have been used previously as part of scraping.

* Convert MetricDefinition implementations

- Updated all the `MetricDefinition` implementations to inherit from `AzureResourceDefinition` instead.

* Update scrapers

I've updated the scraper implementations to work with the new MetricDefinition format. I couldn't come up with a neat way of making MetricDefinition generic without ending up with a nightmare, so for now what I've done is altered the signature of `ScrapeResourceAsync` to accept both the `ScrapeDefinition` object, as well as the `AzureResourceDefinition` cast to the correct type for the subclass.

* Update plumbing code

Updated a couple of the objects used to configure the scraping to handle multiple resources per metric.

* Update validation objects

- Removed the `MetricValidator` abstract class because it doesn't add anything now that we can't cast the `MetricDefinition` to a specific resource type.
- Updated the validators to validate all the resources in the metric definition.

* Update unit tests to handle new MetricDefinition

- Updating `MetricsDeclarationBuilder` to use v1 Builder objects instead of the model objects.
- Updating the tests to handle the fact that a `MetricDefinition` isn't a resource anymore.

* Rename AzureResourceDefinition subclasses

When I was making the initial set of changes, I didn't want to make too many changes at once, so I held off on renaming the subclasses of `AzureResourceDefinition`. This change renames them all from `{Type}MetricDefinition` to `{Type}ResourceDefinition` to make it clearer what they are.

* Populate MetricsDeclarationBuilder.Version

I realised that the Version property wasn't being populated on either the model object or the serialization object. I've removed it from the model object since it isn't versioned.

I've also updated the v1 ConfigurationSerializer to just set the version to `v1` since it only handles the v1 format.

* fixup! Populate MetricsDeclarationBuilder.Version

* fixup! Populate MetricsDeclarationBuilder.Version