Fix markdownlint issues in files
I've tried to only fix the issues highlighted by markdownlint in this commit, and keep the files as close as possible to how they were before these changes, although I did fix a couple of typos that I spotted as I was going through.

I've had to increase the heading level by one on most of the files because they had multiple h1s.