Converted AzureMetricConfigurationDeserializer
Converted AzureMetricConfigurationDeserializer to the new way of defining a deserializer, and added support in the base Deserializer for using another deserializer to deserialize child properties.

I had to add a new `IDeserializer` interface that isn't generic to allow us to deserialize a child property at runtime without knowing its type.

Part of #592