Don't reuse context when deserializing
Because the deserializers are singletons, the set of FieldContext objects was being reused when deserializing multiple nodes of the same type. This lead to a bug where if we had multiple metric definitions, and the first one to be deserialized set a required field, an error would not be reported if that required field was missing from subsequent definitions.

To solve this I've pulled the FieldContext class out of the Deserializer, and I've split it into two separate objects:

- `FieldDeserializationInfo` - contains immutable information about the field to be deserialized, and can be reused between multiple deserializations.
- `FieldDeserializationContext` - contains any information about a specific deserialization attempt, and must be created each time we deserialize a node using a deserializer.

In addition, I've also altered the way the errors / warnings are output to avoid issues I was seeing where the messages weren't always output in the correct order because of the way the logger works.

Part of #592