Fix MD027 violation on sql-server.md
markdownlint was complaining because there was a blockquote defined without an angle bracket on each line. This triggered <https://github.com/DavidAnson/markdownlint/blob/master/doc/Rules.md#md027---multiple-spaces-after-blockquote-symbol>.

I think what's happened is that the docs folder wasn't configured to trigger an Azure Pipelines build, so the check was never run on the PR. I've updated the pipelines file to trigger a build on changes to the docs folder now.