Refactor deserializer config to use builder
I've refactored the way the fields to deserialize are built up to use a builder pattern. The reason for this is so that we can simplify the mapping methods in the `Deserializer` class down to a single `Map()` entry point, and then use separate methods to configure different aspects of the mapping.

The reason I want to do this is in preparation of adding the ability to specify custom validation for fields, for example to validate that Cron expressions are in a valid format. Otherwise we'd end up having to add a load of extra overloads.