Add a tour for creating a new Scraper
I've added a CodeTour explaining how to add a new scraper. I also renamed the `C# Project Overview` tour to `Project Overview` because the `#` character was causing issues with linking to the tour, and I added a link to both tours to the end of the repo overview tour.

Part of #1111.